use /^u|g$/ to check document.readyState. closes #2.
as jdalton suggests, if we're going to rely on the value
of readyState, then we should include `interactive`
as well. i wanted to avoid a RegExp, but this one should
be incredibly fast, and also allows us to get rid of the
original guard to guarantee that the value exists.